requirements.txt
modal
fastapi
pydantic
supabase
psycopg2-binary
python-dateutil
openai>=1.12.0
tiktoken>=0.5.0
